refactor: modularize weather info display
